wezley project structure
