--- revaBooks/src/main/java/com/revature/Genre.java
@@ -1,4 +1,0 @@
-package com.revature;
-
-public enum Genre {
-}
