--- revaBooks/src/main/java/com/revature/repos/BookRepository.java
@@ -1,4 +1,0 @@
-package com.revature.repos;
-
-public class BookRepository {
-}
