--- revaBooks/src/main/java/com/revature/repos/CrudRepository.java
@@ -1,4 +1,0 @@
-package com.revature.repos;
-
-public interface CrudRepository {
-}
