--- revaBooks/src/main/java/com/revature/repos/UserRepository.java
@@ -1,4 +1,0 @@
-package com.revature.repos;
-
-public class UserRepository {
-}
