+++ revaBooks/src/main/java/com/revature/revabooks/models/Genre.java
@@ -1,0 +1,4 @@
+package com.revature.revabooks.models;
+
+public enum Genre {
+}
