+++ revaBooks/src/main/java/com/revature/revabooks/models/Role.java
@@ -1,4 +1,4 @@
-package com.revature;
+package com.revature.revabooks.models;
 
 public enum Role {
     // String roleName;
