+++ revaBooks/src/main/java/com/revature/revabooks/repos/CrudRepository.java
@@ -1,0 +1,4 @@
+package com.revature.revabooks.repos;
+
+public interface CrudRepository {
+}
