+++ revaBooks/src/main/java/com/revature/revabooks/repos/UserRepository.java
@@ -1,0 +1,4 @@
+package com.revature.revabooks.repos;
+
+public class UserRepository {
+}
