+++ revaBooks/src/main/java/com/revature/revabooks/screens/LoginScreen.java
@@ -1,0 +1,4 @@
+package com.revature.revabooks.screens;
+
+public class LoginScreen {
+}
