+++ revaBooks/src/main/java/com/revature/revabooks/services/UserService.java
@@ -1,0 +1,4 @@
+package com.revature.revabooks.services;
+
+public class UserService {
+}
