--- revaBooks/src/main/java/com/revature/screens/DashboardScreen.java
@@ -1,4 +1,0 @@
-package com.revature.screens;
-
-public class DashboardScreen {
-}
