--- revaBooks/src/main/java/com/revature/screens/Screen.java
@@ -1,4 +1,0 @@
-package com.revature.screens;
-
-public abstract class Screen {
-}
