--- revaBooks/src/main/java/com/revature/screens/SearchBooksScreen.java
@@ -1,4 +1,0 @@
-package com.revature.screens;
-
-public class SearchBooksScreen {
-}
