--- revaBooks/src/main/java/com/revature/services/UserService.java
@@ -1,4 +1,0 @@
-package com.revature.services;
-
-public class UserService {
-}
